+++ src/optimization/vqee/vqee.cpp
@@ -65,6 +65,181 @@ namespace qb::vqee {
     return observable;
   }
 
+  size_t VQEE::getOptimumIterationE() {
+    vqe_iteration_data xacc_opt;
+    xacc_opt.energy = params_.optimalValue;
+    xacc_opt.params = params_.theta;
+
+    auto optimum_iteration = std::find(params_.iterationData.begin(), params_.iterationData.end(), xacc_opt);
+    if (optimum_iteration == params_.iterationData.end()) {
+      throw std::range_error("Could not find the iteration where the optimum occurred");
+    } else {
+      size_t index = std::distance(params_.iterationData.begin(), optimum_iteration);
+      return index;
+    }
+  }
+ 
+  void VQEE::generateThetaEnergyVis(const std::string in_title, const size_t in_start_elem, const int in_scale, const int in_width, const int in_precision) {
+    double scale_factor_theta = in_scale/(2*M_PI);
+    std::stringstream ss;    
+    if (!params_.plain) ss << "\033[0m"; // Switch text colour to default
+    ss << "\n";
+    int iterations = params_.iterationData.size();
+    size_t iteration_optimum_e = getOptimumIterationE();    
+    double energy_at_optimum = params_.iterationData.at(iteration_optimum_e).energy;
+
+    size_t use_last = (params_.tail > 0) ? (iterations-params_.tail) : 0;
+    if (params_.blocked) {
+      int iteration_i = use_last;
+      for (size_t i = use_last; i < params_.iterationData.size(); i++) {
+        iteration_i++; // the first iteration is at index 1    
+        if ((params_.iterationData.at(i)).energy == energy_at_optimum) {
+          if (!params_.plain) ss << "\033[1;32m"; // Switch text colour to green bold
+        }     
+        if (iteration_i == (iteration_optimum_e + 1)) {
+          if (!params_.plain) ss << "\033[1;31m"; // Switch text colour to red bold
+        }
+        ss << "Iteration " << iteration_i << "\n";
+
+        // Output the energy for iteration_i
+        double scale_factor = in_scale/std::abs(params_.iterationData.at(0).energy);
+        int bar_scale = 2*in_scale;
+        int bar_val = std::floor(params_.iterationData.at(i).energy * scale_factor);
+        if ((bar_scale + bar_val) < 0) {
+          bar_scale*=2;
+          ss << "Rescaling bars..." << "\n";
+        }
+        ss << "Energy      ";
+        if (iteration_i == (iteration_optimum_e+1)) {
+          ss << std::setw(in_width) << " **|";
+        } else {
+          ss << std::setw(in_width) << "   |";
+        }
+        for (size_t j = 0; j < (bar_scale + bar_val); ++j) {
+          ss << "#";
+        }
+        ss << " " << std::setprecision(in_precision) << (params_.iterationData.at(i)).energy << "\n"; 
+
+        // Output the theta elements for the iteration_i
+        if (params_.showTheta) {
+          size_t use_elements_n = (params_.limitThetaN > 0) ? params_.limitThetaN : params_.theta.size();
+          if (use_elements_n > params_.theta.size()) {
+            throw std::range_error("Requested number of elements exceeds the limit of theta elements");
+          }
+          if (use_elements_n > 0)
+            ss << "Theta"
+               << "\n";
+          for (size_t subpl = in_start_elem; subpl < use_elements_n; ++subpl) {
+            double current_val = (params_.iterationData.at(i).params).at(subpl);
+            int bar_val = std::floor(current_val * scale_factor_theta);
+            ss << std::left << "       Element " << std::setw(4) << subpl;
+            if (iteration_i == (iteration_optimum_e + 1)) {
+              ss << std::setw(in_width) << " **|";
+            }
+            else {
+              ss << std::setw(in_width) << "   |";
+            }
+            for (size_t j = 0; j < (in_scale + bar_val); ++j) {
+              ss << "#";
+            }
+            ss << " " << std::setprecision(in_precision) << (current_val / M_PI) << "*pi"
+               << "\n";
+          }
+        }
+        if (!params_.plain) ss << "\033[0m"; // Switch text colour to default
+      }
+      if (!params_.plain) ss << "\033[0m"; // Switch text colour to default
+    }
+    else {
+      if (params_.showTheta) {
+        size_t use_elements_n = (params_.limitThetaN > 0) ? params_.limitThetaN : params_.theta.size();
+        if (use_elements_n > params_.theta.size()) {
+          throw std::range_error("Requested number of elements exceeds the limit of theta elements");
+        }
+        for (size_t subpl = in_start_elem; subpl < use_elements_n; ++subpl) {
+          ss << in_title << ", Element " << subpl << ", " << iterations << " iterations\n";
+          int iteration_i = use_last;
+          //
+          // Generate trace for theta
+          for (size_t i = use_last; i < params_.iterationData.size(); i++) {
+            iteration_i++; // the first iteration is at index 1
+            if (iteration_i == (iteration_optimum_e + 1)) {
+              if (!params_.plain)
+                ss << "\033[1;31m"; // Switch text colour to red bold
+            }
+            double current_val = (params_.iterationData.at(i).params).at(subpl);
+            int bar_val = std::floor(current_val * scale_factor_theta);
+            ss << std::left << "Iteration " << std::setw(4) << iteration_i;
+            if (iteration_i == (iteration_optimum_e + 1)) {
+              ss << std::setw(in_width) << " **|";
+            }
+            else {
+              ss << std::setw(in_width) << "   |";
+            }
+            for (size_t j = 0; j < (in_scale + bar_val); ++j) {
+              ss << "#";
+            }
+            ss << " " << std::setprecision(in_precision) << (current_val / M_PI) << "*pi"
+               << "\n";
+            if (!params_.plain)
+              ss << "\033[0m"; // Switch text colour to default
+          }
+          if (!params_.plain)
+            ss << "\033[0m"; // Switch text colour to default
+        }
+      }
+      if (!params_.plain) ss << "\033[0m"; // Switch text colour to default    
+      ss << generateEnergyVis(params_.energies, "Energy");
+    }
+    if (!params_.plain) ss << "\033[0m"; // Switch text colour to default
+    params_.vis = ss.str();
+  }
+
+  std::string VQEE::generateEnergyVis(const std::vector<double> in_val, const std::string in_title, const int in_stride, const int in_scale,  const int in_width, const int in_precision) {
+    double scale_factor = in_scale/std::abs(in_val.at(0));
+    int bar_scale = 2*in_scale;
+    std::stringstream ss;
+    if (!params_.plain) ss << "\033[0m"; // Switch text colour to default
+
+    int iterations = in_val.size()/in_stride;    
+    size_t iteration_optimum = getOptimumIterationE();
+    double energy_at_optimum = in_val.at(iteration_optimum);
+    size_t use_last = (params_.tail > 0) ? (iterations-params_.tail) : 0;
+
+    for (size_t subpl = 0; subpl < in_stride; ++subpl) {        
+      ss << in_title << " element " << subpl << ", " << iterations << " iterations\n";
+      size_t iteration_i = use_last;
+      for (size_t i = subpl+use_last*in_stride; i < in_val.size(); i += in_stride) {
+        iteration_i++;  // the first iteration is at index 1        
+        int bar_val = std::floor(in_val.at(i)*scale_factor);
+        if ((bar_scale + bar_val) < 0) {
+          bar_scale*=2;
+          ss << "Rescaling bars..." << "\n";
+        }       
+        if (in_val.at(i) == energy_at_optimum) {
+          if (!params_.plain) ss << "\033[1;32m"; // Switch text colour to green bold
+        }       
+        if (iteration_i == (iteration_optimum+1)) {
+          if (!params_.plain) ss << "\033[1;31m"; // Switch text colour to red bold
+        }
+        ss << std::left << "Iteration " << std::setw(4) << iteration_i;
+        if (iteration_i == (iteration_optimum+1)) {
+          ss << std::setw(in_width) << " **|";
+        } else {
+          ss << std::setw(in_width) << "   |";
+        }
+        for (size_t j = 0; j < (bar_scale + bar_val); ++j) {
+          ss << "#";
+        }
+        ss << " " << std::setprecision(in_precision) << in_val.at(i) << "\n"; 
+        if (!params_.plain) ss << "\033[0m"; // Switch text colour to default
+      }
+      if (!params_.plain) ss << "\033[0m"; // Switch text colour to default
+    }    
+    if (!params_.plain) ss << "\033[0m"; // Switch text colour to default
+    return ss.str();
+  }
+
   void VQEE::optimize(){ 
   // Here VQE is called with a decorated accelerator. The decorator adds pre- and post-processing around the actual accelerator execution. 
   // This is used to introduce MPI parallelism, i.e partitioning and distributing the vector of instructions (base curcuit + Pauli terms) 
@@ -98,6 +273,25 @@ namespace qb::vqee {
       params_.energies       = (*buffer)["params-energy" ].as<std::vector<double>>();
       params_.theta          = (*buffer)["opt-params"    ].as<std::vector<double>>();
       params_.optimalValue   = (*buffer)["opt-val"       ].as<double>();
+      
+      const size_t nIters = params_.energies.size();
+      const size_t step = (buffer->nChildren()) / nIters;
+      size_t stepidx = 0;
+
+      // params_.iterationData
+      params_.iterationData.clear();
+      for (auto &childBuff : buffer->getChildren()) {
+        if (stepidx % step == 0) {
+          if (childBuff->hasExtraInfoKey("parameters")) {
+            std::vector<double> param = (*childBuff)["parameters"].as<std::vector<double>>();            
+            vqe_iteration_data vid;
+            vid.energy = params_.energies.at(stepidx/step);
+            vid.params = param;
+            params_.iterationData.emplace_back(vid);
+          }
+        }
+        stepidx++;
+      }
     } 
     else {
       const int nOptVars = ansatz->nVariables();
@@ -152,6 +346,9 @@ namespace qb::vqee {
     }
 
     if (isRoot_) {
+      if (params_.enableVis) {
+        generateThetaEnergyVis("theta");
+      }
       std::cout << "\nMin energy = "       << params_.optimalValue << "\n"
                 << "Optimal parameters = " << params_.theta        << std::endl;
     }